requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.8
django-classy-tags==4.1.0
django-js-asset==2.2.0
django-mptt==0.14.0
django-nyt==1.3
django-sekizai==4.1.0
filelock==3.13.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-aiplatform==1.38.1
google-cloud-bigquery==3.14.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.11.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.3.1
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
langchain-core==0.1.1
langchain-google-genai==0.0.5
langsmith==0.0.72
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
multi-rake==0.0.2
networkx==3.2.1
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
proto-plus==1.23.0
protobuf==4.25.1
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycld2==0.41
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
pyrsistent==0.20.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rpds-py==0.14.1
rsa==4.9
safetensors==0.4.1
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sorl-thumbnail==12.10.0
soupsieve==2.5
sqlparse==0.4.4
streamlit==1.29.0
streamlit-searchbox ==0.1.6
sympy==1.12
tenacity==8.2.3
tinycss2==1.2.1
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
webencodings==0.5.1
wiki==0.10
wikipedia==1.4.0
zipp==3.17.0
google-cloud-discoveryengine